app/requirements.txt
fastapi==0.58.1
python-dotenv==0.19.0
uvicorn==0.15.0
requests==2.26.0
opentelemetry-distro==0.38b0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
